% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biospheremetrics-package.R
\docType{package}
\name{biospheremetrics-package}
\alias{biospheremetrics}
\alias{biospheremetrics-package}
\title{biospheremetrics: Biosphere integrity metrics for LPJmL}
\description{
Functions to compute Biosphere integrity metrics BioCol and EcoRisk based on output from LPJmL.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stenzelf/biospheremetrics}
}

}
\author{
\strong{Maintainer}: Fabian Stenzel \email{stenzel@pik-potsdam.de} (\href{https://orcid.org/0000-0002-5109-0048}{ORCID})

Authors:
\itemize{
  \item Jannes Breier \email{jannesbr@pik-potsdam.de} (\href{https://orcid.org/0000-0002-9055-6904}{ORCID})
  \item Johanna Braun \email{jobraun@pik-potsdam.de} (\href{https://orcid.org/0000-0002-8809-1044}{ORCID})
}

}
\keyword{internal}
