% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECO.R
\name{calcDeltaV}
\alias{calcDeltaV}
\title{Calculates changes in vegetation structure (deltaV)}
\usage{
calcDeltaV(
  fpcRef,
  fpcScen,
  bftRef,
  bftScen,
  cftRef,
  cftScen,
  weighting = "equal"
)
}
\arguments{
\item{fpcRef}{reference fpc array (dim: [ncells,npfts+1])}

\item{fpcScen}{scenario fpc array (dim: [ncells,npfts+1])}

\item{bftRef}{reference bft array (dim: [ncells,nbfts])}

\item{bftScen}{scenario bft array (dim: [ncells,nbfts])}

\item{cftRef}{reference cft array (dim: [ncells,ncfts])}

\item{cftScen}{scenario cft array (dim: [ncells,ncfts])}

\item{weighting}{apply "old" (Ostberg-like), "new", or "equal" weighting of deltaV weights (default "equal")}
}
\value{
deltaV array of size ncells with the deltaV value [0,1] for each cell
}
\description{
Utility function to calculate changes in vegetation structure (deltaV)
for calculation of MECO
}
\examples{
\dontrun{
deltaV <- calcDeltaV(fpcRef = fpc_ref_mean,fpcScen = apply(fpc_scen,c(1,2),mean),
          bftRef = bft_ref_mean,bftScen = apply(bft_scen,c(1,2),mean),
          cftRef = cft_ref_mean, cftScen = apply(cft_scen,c(1,2),mean),
          weighting = "equal")
}
}
