% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECO.R
\name{calcGamma}
\alias{calcGamma}
\title{Calculate the ecosystem change metric gamma between 2 simulations/timesteps}
\usage{
calcGamma(
  folderRef,
  folderRef2 = NULL,
  folderScen,
  readPreviouslySavedData = TRUE,
  saveFileData = NULL,
  saveFileGamma = NULL,
  combined = FALSE,
  nitrogen = FALSE,
  weighting = "old",
  varnames = NULL,
  headerout = 0,
  timespan_full_ref,
  timespan_full_scen,
  timespan_focus_ref,
  timespan_focus_ref2 = NULL,
  timespan_focus_scen,
  npfts,
  ncfts,
  nbfts,
  ncells,
  dimensionsOnlyLocal = T
)
}
\arguments{
\item{folderRef}{folder of reference run}

\item{folderRef2}{2nd folder of reference run, if split (default NULL -- no split)}

\item{folderScen}{folder of scenario run}

\item{saveFileData}{file to save read in data to (default NULL)}

\item{saveFileGamma}{file to save gamma data to (default NULL)}

\item{combined}{read in soilc+litc, evap+interc, rh+harvestc separate or as combined outputs (default F -- separate)}

\item{nitrogen}{include nitrogen outputs for pools and fluxes into gamma calculation (default F)}

\item{weighting}{apply "old" (Ostberg-like), "new", or "equal" weighting of deltaV weights (default "old")}

\item{varnames}{data.frame with names of output files -- can be specified to account for variable file names (default NULL -- standard names are used)}

\item{headerout}{headersize of the output files (default 0)}

\item{timespan_full_ref}{full timespan of output files in reference period}

\item{timespan_full_scen}{full timespan of output files in scenario period}

\item{timespan_focus_ref}{specific years to use as reference period}

\item{timespan_focus_ref2}{specific years to use as 2nd part of reference period in case this is split (default NULL)}

\item{timespan_focus_scen}{specific years to use as scenario period}

\item{npfts}{number of natural plant functional types (== bands in fpc - 1)}

\item{ncfts}{number of crop functional types}

\item{nbfts}{number of bfts (bioenergy functional types)}

\item{ncells}{number of cells in lpjml grid}

\item{exportGamma}{flag whether to export gamma data to global environment (default F)}
}
\value{
list data object containing arrays of deltaV, local_change,
        global_change, ecosystem_balance, carbon_fluxes, carbon_stocks,
        water_fluxes, gamma_total (+ nitrogen_fluxes and nitrogen_stocks if used)
}
\description{
Function to calculate the ecosystem change metric gamma according
to Sykes (1999), Heyder (2011), and Ostberg (2015,2018).
This is a reformulated version in R, not producing 100% similar values
than the C/bash version from Ostberg 2018, but following their methodology.
}
\examples{
\dontrun{
gamma_today <- calcGamma(folderRef = "/p/projects/elis/ostberg/biome_shift/archive_Ostberg2018/simulations/scenario_PNV_trunkupdate_noCC/HadGEM2-ES/r1i1p1/",
               folderScen = "/p/projects/elis/ostberg/biome_shift/archive_Ostberg2018/simulations/scenario_LUC_trunkupdate/rcp26/HadGEM2-ES/r1i1p1/luc_rcp26_50bg50btflexi/",
               readPreviouslySavedData = TRUE, saveFileData = "/media/All/Gamma/Gamma_today_data.RData",
               saveFileGamma = "/media/All/Gamma/Gamma_today_gamma.RData",
               timespan_full_ref = c(1991,2099), timespan_full_scen = c(1991,2099),
               timespan_focus_ref = c(2000,2029), timespan_focus_scen = c(2000,2029),
               npfts = 9, ncfts = 32, nbfts = 12, ncells = 64240,
               combined = TRUE, weighting = "old")
}
}
