% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECO.R
\name{calcMECO}
\alias{calcMECO}
\title{Calculate the ecosystem change metric MECO between 2 simulations/timesteps}
\usage{
calcMECO(
  folderRef,
  folderRef2 = NULL,
  folderScen,
  readPreviouslySavedData = FALSE,
  saveFileData = NULL,
  saveFileMECO = NULL,
  combined = FALSE,
  nitrogen = FALSE,
  weighting = "old",
  varnames = NULL,
  headerout = 0,
  timespan_full_ref,
  timespan_full_scen,
  timespan_focus_ref,
  timespan_focus_ref2 = NULL,
  timespan_focus_scen,
  npfts,
  ncfts,
  nbfts,
  ncells,
  soillayers = 6,
  dimensionsOnlyLocal = F
)
}
\arguments{
\item{folderRef}{folder of reference run}

\item{folderRef2}{optional 2nd reference run folder (default NULL - no split)}

\item{folderScen}{folder of scenario run}

\item{readPreviouslySavedData}{whether to read in previously saved data 
(default: FALSE)}

\item{saveFileData}{file to save read in data to (default NULL)}

\item{saveFileMECO}{file to save MECO data to (default NULL)}

\item{combined}{read in soilc+litc, evap+interc, rh+harvestc separate or as 
combined outputs (default F -- separate)}

\item{nitrogen}{include nitrogen outputs for pools and fluxes into MECO 
calculation (default F)}

\item{weighting}{apply "old" (Ostberg-like), "new", or "equal" weighting of 
deltaV weights (default "old")}

\item{varnames}{data.frame with names of output files (outname) and time res.
(timestep) -- can be specified to account for variable file names 
(default NULL -- standard names as below)
varnames <- data.frame(row.names = c("grid",    "fpc",    "fpc_bft",    "cftfrac",    "firec",    "rh_harvest",    "npp",    "evapinterc",    "runoff",    "transp",    "soillitc",    "vegc",    "swcsum",    "firef",    "rh",    "harvestc",    "evap",    "interc",    "soilc",    "litc",    "swc",    "vegn",    "soilnh4",    "soilno3",    "leaching",    "n2o_denit",    "n2o_nit",    "n2o_denit",    "n2_emis",    "bnf",    "n_volatilization"),
outname = c("grid.bin","fpc.bin","fpc_bft.bin","cftfrac.bin","firec.bin","rh_harvest.bin","npp.bin","evapinterc.bin","runoff.bin","transp.bin","soillitc.bin","vegc.bin","swcsum.bin","firef.bin","rh.bin","harvestc.bin","evap.bin","interc.bin","soilc.bin","litc.bin","swc.bin","vegn.bin","soilnh4.bin","soilno3.bin","leaching.bin","n2o_denit.bin","n2o_nit.bin","n2o_denit.bin","n2_emis.bin","bnf.bin","n_volatilization.bin"),
timestep = c("Y",       "Y",      "Y",          "Y",          "Y",        "Y",             "Y",      "Y",             "Y",         "Y",         "Y",           "Y",       "Y",         "Y",        "Y",     "Y",           "Y",       "Y",         "Y",        "Y",       "Y",      ,"Y",      "Y",          "Y",          "Y",           "Y",            "Y",          "Y",            "Y",          "Y",      "Y"))}

\item{headerout}{headersize of the output files (default: 0 bytes)}

\item{timespan_full_ref}{full timespan of output files in reference period}

\item{timespan_full_scen}{full timespan of output files in scenario period}

\item{timespan_focus_ref}{specific years to use as reference period}

\item{timespan_focus_ref2}{specific years to use as 2nd part of reference 
period in case this is split (default NULL)}

\item{timespan_focus_scen}{specific years to use as scenario period}

\item{npfts}{number of natural plant functional types (== bands in fpc - 1)}

\item{ncfts}{number of crop bands (rainfed+irrigated)}

\item{nbfts}{number of bands in fpc_bft (bioenergy functional types)}

\item{ncells}{number of cells in lpjml grid}

\item{soillayers}{number of soil layers (default: 6)}

\item{dimensionsOnlyLocal}{flag whether to use only local change component
for water/carbon/nitrogen fluxes and pools, or use an average of
local change, global change and ecosystem balance (default F)}
}
\value{
list data object containing arrays of deltaV, local_change,
        global_change, ecosystem_balance, carbon_fluxes, carbon_stocks,
        water_fluxes, meco_total (+ nitrogen_fluxes and nitrogen_stocks)
}
\description{
Function to calculate the ecosystem change metric MECO, based on gamma/deltaV
work from Sykes (1999), Heyder (2011), and Ostberg (2015,2018).
This is a reformulated version in R, not producing 100% similar values
than the C/bash version from Ostberg et al. 2018, but similar the methodology
}
\examples{
\dontrun{
}
}
