% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_biomes.R
\name{classify_biomes}
\alias{classify_biomes}
\title{Classify biomes}
\usage{
classify_biomes(
  path_data,
  timespan,
  input_files = NULL,
  diff_output_files = NULL,
  file_type = "raw",
  savanna_proxy = list(pft_lai = 6),
  montane_arctic_proxy = list(elevation = 1000),
  tree_cover_thresholds = list(),
  avg_nyear_args = list(),
  read_args = list(header_size = 0, ncell = 67420, firstyear = 1901, fpc_nbands = 12,
    size = 4)
)
}
\arguments{
\item{path_data}{output directory (character string) of the LPJmL run o read
default outputs from (fpc, grid, vegc, pft_lai, temp)}

\item{timespan}{time span to be used, defined as an integer vector, e.g.
`c(1982,2011)`, to average over (default over all years, else see
\link[pbfunctions]{average_nyear_window})}

\item{input_files}{optional `list` containing additional input (!) files,
not in the `path_data`, e.g. if temp was not written out:
`list(grid=..., temp = ..., elevation = ...)`}

\item{diff_output_files}{optional list for specification of output file names
differing from default, which is list(grid = "grid.bin", fpc = "fpc.bin",# nolint
vegc = "vegc.bin", pft_lai = "pft_lai.bin", temp = "temp.bin")}

\item{savanna_proxy}{`list` with either "pft_lai" or "vegc" as
name/key and value in m2/m2 for pft_lai (default = 6) and gC/m2 for
vegc (default would be 7500), Set to `NULL` if no proxy is used.}

\item{montane_arctic_proxy}{`list` with either "elevation" or "latitude" as
name/key and value in m for elevation (default 1000) and degree for
latitude (default would be 55), Set to `NULL` if no proxy is used.}

\item{tree_cover_thresholds}{list with minimum tree cover thresholds for
definition of forest, woodland, savanna and grassland. Only changes to
the default have to be included in the list, for the rest the default
is used. Default values:
"boreal forest" = 0.6
"temperate forest" = 0.6
"temperate woodland" = 0.3
"temperate savanna" = 0.1
"tropical forest" = 0.6
"tropical woodland" = 0.3
"tropical savanna" = 0.1}

\item{avg_nyear_args}{list of arguments to be passed to
\link[pbfunctions]{average_nyear_window} (see for more info). To be used for # nolint
time series analysis}

\item{read_args}{list of arguments for reading input/output. only required
for:
nc output: specification of header_size and ncell to read in
           lpjml grid input
raw/clm output: specification of header_size, ncell, firstyear and 
           fpc_nbands (12 or 10)}

\item{file_ending}{replace default file ending. default: ".bin"}
}
\value{
list object containing biome_id (main biome per grid cell [dim=c(ncells)]), # nolint
and list of respective biome_names[dim=c(nbiomes)]
}
\description{
Classify biomes based on foliage protected cover (FPC) and temperature
LPJmL output plus either vegetation carbon or pft_lai depending on
the savanna_proxy option and elevation if montane_arctic_proxy requires this
}
\examples{
\dontrun{
classify_biomes(
  path_data = "/p/projects/open/Fabian/runs/Gamma/output/historic_gamma"
  timespan = c(1982,2011))
}

}
