% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECO.R
\name{disaggregateGammaIntoBiomes}
\alias{disaggregateGammaIntoBiomes}
\title{Returns the average gamma values across regions}
\usage{
disaggregateGammaIntoBiomes(
  gamma,
  biome_class,
  type = "quantile",
  classes = "4biomes"
)
}
\arguments{
\item{gamma}{gamma data object with dimension c(regions/biomes[4/16],gamma_components[10/8],minmeanmax/quantiles[3])}

\item{biome_class}{path to save modified data to}

\item{type}{return minimum,mean,maximum or Q10,Q50,Q90}

\item{classes}{into which regions should be disaggregated? "4biomes" or "allbiomes" default "4biomes"}
}
\description{
Returns the average value across either 4 regions or 19 biomes for gamma
and each of the subcomponents for each
}
\examples{
\dontrun{
disaggregateGammaIntoBiomes(gamma = gamma_historic_nitrogen_classic_bnf,
                   biome_class = gamma_historic_withNitrogen_biome_classes,
                   type = "quantile",classes = "4biomes")
}
}
