% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECO.R
\name{disaggregateMECOintoBiomes}
\alias{disaggregateMECOintoBiomes}
\title{Averages MECO values across regions}
\usage{
disaggregateMECOintoBiomes(
  meco,
  biome_class,
  type = "quantile",
  classes = "4biomes"
)
}
\arguments{
\item{meco}{MECO list object as returned by calcMECO}

\item{biome_class}{biome class list object as returned by classify_biomes}

\item{type}{string controlling whether to return  minimum,mean,maximum 
("minmeanmax") or Q10,Q50,Q90 ("quantile") - default: "quantile"}

\item{classes}{string for into how many regions should be disaggregated 
"4biomes" (tropics/temperate/boreal/arctic) or "allbiomes"}
}
\description{
Returns the average value across either 4 regions or all (19) biomes for MECO
and each of the subcomponents for each
}
\examples{
\dontrun{
disaggregateMECOintoBiomes(meco = meco,
                   biome_class = biome_classes,
                   type = "quantile",classes = "4biomes")
}
}
