% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecorisk.R
\name{ecorisk_combine_hist_and_scen_data}
\alias{ecorisk_combine_hist_and_scen_data}
\title{Create modified EcoRisk data combining two time series}
\usage{
ecorisk_combine_hist_and_scen_data(hist_file, scen_file, combined_file)
}
\arguments{
\item{hist_file}{path to input file with historic data}

\item{scen_file}{path to input file with scenario data}

\item{combined_file}{path to save modified data to}
}
\description{
Function to combine two EcoRisk data files (e.g. historic and futures) into
one file.
}
