% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECO.R
\name{ecoriskWrapper}
\alias{ecoriskWrapper}
\title{Wrapper for calculating the ecosystem change metric EcoRisk}
\usage{
ecoriskWrapper(
  folderRef,
  folderScen,
  readPreviouslySavedData = FALSE,
  saveFileData = NULL,
  saveFileEcoRisk = NULL,
  nitrogen = TRUE,
  weighting = "equal",
  varnames = NULL,
  time_span_reference,
  time_span_scenario,
  dimensionsOnlyLocal = F,
  overtime = F,
  window = 30,
  debug = FALSE
)
}
\arguments{
\item{folderRef}{folder of reference run}

\item{folderScen}{folder of scenario run}

\item{readPreviouslySavedData}{whether to read in previously saved data
(default: FALSE)}

\item{saveFileData}{file to save read in data to (default NULL)}

\item{saveFileEcoRisk}{file to save EcoRisk data to (default NULL)}

\item{nitrogen}{include nitrogen outputs for pools and fluxes into EcoRisk
calculation (default F)}

\item{weighting}{apply "old" (Ostberg-like), "new", or "equal" weighting of
vegetation_structure_change weights (default "equal")}

\item{varnames}{data.frame with names of output files (outname) and time res.
(timestep) -- can be specified to account for variable file names
(default NULL -- standard names as below)
varnames <- data.frame(row.names = c("grid",    "fpc",    "fpc_bft",    "cftfrac",    "firec",    "rh_harvest",    "npp",    "evapinterc",    "runoff",    "transp",    "soillitc",    "vegc",    "swcsum",    "firef",    "rh",    "harvestc",    "evap",    "interc",    "soilc",    "litc",    "swc",    "vegn",    "soilnh4",    "soilno3",    "leaching",    "n2o_denit",    "n2o_nit",    "n2o_denit",    "n2_emis",    "bnf",    "n_volatilization"),
outname = c("grid.bin","fpc.bin","fpc_bft.bin","cftfrac.bin","firec.bin","rh_harvest.bin","npp.bin","evapinterc.bin","runoff.bin","transp.bin","soillitc.bin","vegc.bin","swcsum.bin","firef.bin","rh.bin","harvestc.bin","evap.bin","interc.bin","soilc.bin","litc.bin","swc.bin","vegn.bin","soilnh4.bin","soilno3.bin","leaching.bin","n2o_denit.bin","n2o_nit.bin","n2o_denit.bin","n2_emis.bin","bnf.bin","n_volatilization.bin"),
timestep = c("Y",       "Y",      "Y",          "Y",          "Y",        "Y",             "Y",      "Y",             "Y",         "Y",         "Y",           "Y",       "Y",         "Y",        "Y",     "Y",           "Y",       "Y",         "Y",        "Y",       "Y",      ,"Y",      "Y",          "Y",          "Y",           "Y",            "Y",          "Y",            "Y",          "Y",      "Y"))}

\item{time_span_reference}{vector of years to use as scenario period}

\item{time_span_scenario}{vector of years to use as scenario period}

\item{dimensionsOnlyLocal}{flag whether to use only local change component
for water/carbon/nitrogen fluxes and pools, or use an average of
local change, global change and ecosystem balance (default F)}

\item{overtime}{logical: calculate ecorisk as time-series? (default: F)}

\item{window}{integer, number of years for window length (default: 30)}

\item{debug}{write out all nitrogen state variables (default F)}
}
\value{
list data object containing arrays of ecorisk_total, vegetation_structure_change, 
        local_change, global_importance, ecosystem_balance, carbon_stocks, 
        carbon_fluxes, water_fluxes (+ nitrogen_stocks and nitrogen_fluxes)
}
\description{
Function to read in data for ecorisk, and call the calculation function once,
if overtime is F, or for each timeslice of length window years, if
overtime is T
}
\examples{
\dontrun{
}
}
