% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECO.R
\name{evaluateGamma}
\alias{evaluateGamma}
\title{Create modified gamma data}
\usage{
evaluateGamma(
  dataFileIn,
  dataFileOut,
  temperature,
  pft_lai,
  elevation,
  lat,
  lon,
  refBiom
)
}
\arguments{
\item{dataFileIn}{path to input data}

\item{dataFileOut}{path to save modified data to}

\item{temperature}{yearly average temperature}

\item{refBiom}{}
}
\description{
Function to create a modified gamma data file where each cell is compared
to the average cell of the precribed reference biome refBiom
}
\examples{
\dontrun{
evaluateGamma(dataFileIn = "/media/All/Gamma/Gamma_today_data.RData",
      dataFileOut = "/media/All/Gamma/Gamma_preind_compared_to_average_TrBD_data.RData",
      temperature = temperature_1716, lat = lat64240, lon = lon64240, refBiom = 4)
}
}
