% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECO.R
\name{gammaCrossTable}
\alias{gammaCrossTable}
\title{Create modified gamma data for crosstable}
\usage{
gammaCrossTable(
  dataFileIn,
  dataFileOut,
  temperature,
  pft_lai,
  elevation,
  pickCells = NULL,
  lat,
  lon
)
}
\arguments{
\item{dataFileIn}{path to input data}

\item{dataFileOut}{path to save modified data to}

\item{temperature}{yearly average temperature}

\item{pft_lai}{yearly pft_lai values for the biome classification}

\item{elevation}{elevation input file for the biome classification}

\item{lon}{}
}
\description{
Function to create a modified gamma data file where for each present day biome
the average cell is compared to the average cell of all other biomes. This
can then be used to compute a crosstable with the average difference between
each of them as in the SI of Ostberg et al. 2013
(Critical impacts of global warming on land ecosystems)
}
\examples{
\dontrun{
gammaCrossTable(dataFileIn = "/media/All/Gamma/Gamma_today_data.RData",
      dataFileOut = "/media/All/Gamma/Gamma_preind_compared_to_average_TrBD_data.RData",
      temperature = temperature_1716, pft_lai = pft_lai_1996, elevation = elevation, lat = lat64240, lon = lon64240, refBiom = 4)
}
}
