% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecorisk.R
\name{get_biome_names}
\alias{get_biome_names}
\title{Get biome names}
\usage{
get_biome_names(biome_name_length = 2)
}
\arguments{
\item{biome_name_length}{integer chose from 1,2,3 for abbreviated, short,
or full biome names}
}
\description{
Returns biome names with variable length (abbreviated, short, or full)
}
