% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_needed_outputs.R
\name{list_needed_outputs}
\alias{list_needed_outputs}
\title{Returns LPJmL outputs required for given metric}
\usage{
list_needed_outputs(metric = "all", with_nitrogen = TRUE)
}
\arguments{
\item{metric}{string/list of strings, containing name of metric to get
required outputs for. Pick from "meco", "mcol", "biome", "all_pbs",
       "pb_n", "pb_w", "pb_b", "pb_lsc", "all" (default)}

\item{withNitrogen}{logical: include nitrogen outputs? default: TRUE}
}
\value{
List of strings with LPJmL variable names
}
\description{
Function to return a list of strings with the LPJmL output names required for
the computation of the metrics M-COL, M-ECO, the biome classification
and/or planetary boundary calculations
}
\examples{
\dontrun{

}
}
