% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECO.R
\name{mecoCrossTable}
\alias{mecoCrossTable}
\title{Create modified MECO data for crosstable}
\usage{
mecoCrossTable(dataFileIn, dataFileOut, biome_classes_in, pickCells = NULL)
}
\arguments{
\item{dataFileIn}{path to input data}

\item{dataFileOut}{path to save modified data to}

\item{biome_classes_in}{biome classes object as returned from classify_biomes}

\item{pickCells}{pick one specific cell as representative for the biome 
instead of computing the average state}
}
\description{
Function to create a modified MECO data file where for each biome
the average scenario cell is compared to the average scenario cell of all 
other biomes. This can then be used to compute a crosstable with the average 
difference between each of them as in the SI of Ostberg et al. 2013
(Critical impacts of global warming on land ecosystems)
}
\examples{
\dontrun{
}
}
