% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_biocol.R
\name{plot_biocol_map}
\alias{plot_biocol_map}
\title{Plot global map of BioCol to file}
\usage{
plot_biocol_map(
  data,
  file,
  title = "",
  legendtitle = "",
  zero_threshold = 0.001,
  eps = FALSE,
  haberllegend = FALSE
)
}
\arguments{
\item{data}{array containing BioCol percentage value for each gridcell}

\item{file}{character string for location/file to save plot to}

\item{title}{character string title for plot}

\item{legendtitle}{character string legend title}

\item{zero_threshold}{smallest value to be distinguished from 0 in legend,
both for negative and positive values (default: 0.1)}

\item{eps}{write eps file instead of PNG (boolean) - (default: FALSE)}

\item{plotyears}{range of years to plot over time}
}
\value{
none
}
\description{
Plot global map of BioCol to file with legend colors similar to
Haberl et al. 2007
}
