% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ecorisk.R
\name{plot_biome_internal_distribution}
\alias{plot_biome_internal_distribution}
\title{Plot distribution of similarity within biomes}
\usage{
plot_biome_internal_distribution(
  data,
  file = NULL,
  biomes_abbrv = NULL,
  scale = 1,
  title = "",
  legendtitle = "",
  eps = FALSE,
  palette = NULL
)
}
\arguments{
\item{data}{data object with distibution - as returned by
calculateWithInBiomeDiffs for each subcategory of ecorisk.
dim: c(biomes,bins)}

\item{file}{to write into, if not supplied (default is NULL) write to screen}

\item{biomes_abbrv}{character vector. abbreviated names of biomes
(defaults to NULL -> extract dimension names from data)}

\item{scale}{scaling factor for distribution. defaults to 1}

\item{title}{character string title for plot, default empty}

\item{legendtitle}{character string legend title, default empty}

\item{eps}{write as eps or png (default: FALSE -> png)}

\item{palette}{color palette to plot EcoRisk with, defaults to the Ostberg
color scheme white-blue-yellow-red}
}
\description{
Function to plot the distribution of similarity within biomes
}
\examples{
\dontrun{
plot_biome_internal_distribution(
  data = biomes,
  file = "./biomes.png"
)
}

}
