% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecorisk.R
\name{plot_biome_internal_distribution_to_screen}
\alias{plot_biome_internal_distribution_to_screen}
\title{Plot distribution of similarity within biomes}
\usage{
plot_biome_internal_distribution_to_screen(
  data,
  biomes_abbrv,
  title = "",
  legendtitle = "",
  scale = 1,
  palette = NULL
)
}
\arguments{
\item{data}{data object with distibution - as returned by
calculateWithInBiomeDiffs for each subcategory of ecorisk.
dim: c(biomes,bins)}

\item{biomes_abbrv}{to mask the focus_biome from}

\item{title}{character string title for plot, default empty}

\item{legendtitle}{character string legend title, default empty}

\item{scale}{scaling factor for distribution. defaults to 1}

\item{palette}{color palette to plot EcoRisk with, defaults to the Ostberg
color scheme white-blue-yellow-red}
}
\value{
None
}
\description{
Function to plot the distribution of similarity within biomes
}
