% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ecorisk.R
\name{plot_biomes_mercator}
\alias{plot_biomes_mercator}
\title{Plot biomes with mercator projection}
\usage{
plot_biomes_mercator(
  biome_ids,
  biome_name_length = 1,
  order_legend = "plants",
  file = NULL,
  title = "",
  title_size = 2,
  leg_yes = TRUE,
  leg_scale = 1,
  eps = FALSE
)
}
\arguments{
\item{biome_ids}{biome id as given by classify_biomes}

\item{biome_name_length}{length of biome names in legend: 1 - abbreviation,
2 - short name, 3 - full biome name}

\item{order_legend}{legend order: either "plants" to first have forests, then
grasslands, then tundra ..., or "zones" to go from north to south
(default: "plants")}

\item{file}{to write into (if not supplied - default NULL - prints to screen)}

\item{title}{character string title for plot, default empty}

\item{title_size}{size of title in cex units (defaukt: 2)}

\item{leg_yes}{whether to plot legend (default: True)}

\item{leg_scale}{size of legend in cex units (default 0.5)}

\item{eps}{write as eps, replacing png in filename (default: True)}
}
\description{
Function to plot biomes to file (or screen) using mercator projection
}
\examples{
\dontrun{
plot_biomes_mercator(
  biome_ids = biome_classification$biome_id,
  file = "./biomes.png"
)
}

}
