% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecorisk.R
\name{plot_eco_riskmap_to_screen}
\alias{plot_eco_riskmap_to_screen}
\title{Plot EcoRisk map to screen}
\usage{
plot_eco_riskmap_to_screen(
  data,
  focus_biome = NULL,
  biome_classes = NULL,
  title = "",
  legendtitle = "",
  title_size = 1,
  leg_yes = TRUE,
  palette = NULL
)
}
\arguments{
\item{data}{folder of reference run}

\item{focus_biome}{highlight the biome with this id and desaturate all other
(default NULL -- no highlight)}

\item{biome_classes}{to mask the focus_biome from}

\item{title}{character string title for plot, default empty}

\item{legendtitle}{character string legend title}

\item{leg_yes}{logical. whether to plot legend or not. defaults to TRUE}

\item{palette}{color palette to plot EcoRisk with, defaults to the Ostberg
color scheme white-blue-yellow-red}

\item{leg_scale}{scaling factor for legend. defaults to 1}
}
\value{
None
}
\description{
Function to plot a global map of EcoRisk values [0-1] per grid cell to screen
}
