% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecorisk.R
\name{plot_ecorisk_cross_table_to_screen}
\alias{plot_ecorisk_cross_table_to_screen}
\title{Plot crosstable showing (dis-)similarity between average biome pixels}
\usage{
plot_ecorisk_cross_table_to_screen(data, lmar = 3, palette = NULL)
}
\arguments{
\item{data}{crosstable data as array with [nbiomes,nbiomes] and row/colnames}

\item{lmar}{left margin for plot in lines (default: 3)}

\item{palette}{color palette to plot EcoRisk with, defaults to the Ostberg
color scheme white-blue-yellow-red}
}
\value{
None
}
\description{
Function to plot a crosstable showing (dis-)similarity between average
biome pixels based on EcoRisk (former gamma) metric from LPJmL simulations
}
