% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecorisk.R
\name{plot_ecorisk_radial}
\alias{plot_ecorisk_radial}
\title{Plot radial EcoRisk plot to file}
\usage{
plot_ecorisk_radial(
  data,
  file,
  title = "",
  leg_yes = TRUE,
  eps = FALSE,
  use_quantile = TRUE
)
}
\arguments{
\item{data}{EcoRisk data array c(4/19[biomes],[nEcoRiskcomponents],
3[min,mean,max])}

\item{file}{to write into}

\item{title}{character string title for plot, default empty}

\item{leg_yes}{logical. whether to plot legend or not. defaults to TRUE}

\item{eps}{write as eps or png}

\item{type}{plot type, 'legend1' for variable and color legend,
'legend2' for value legend, or 'regular' (default setting)
for the regular EcoRisk plot}
}
\value{
None
}
\description{
Function to plot an aggregated radial status of EcoRisk values [0-1]
for the different sub-categories to file
}
