% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ecorisk.R
\name{plot_ecorisk_radial_panel}
\alias{plot_ecorisk_radial_panel}
\title{Plot radial EcoRisk panel to file with 4/16 biomes}
\usage{
plot_ecorisk_radial_panel(
  data,
  biome_names,
  file,
  use_quantile = TRUE,
  eps = FALSE
)
}
\arguments{
\item{data}{EcoRisk data array c(4/19[biomes],[nEcoRiskcomponents],
3[min,mean,max])}

\item{biome_names}{names of biomes}

\item{file}{to write into}

\item{use_quantile}{is it quantiles or minmeanmax data? - text for whiskers}

\item{eps}{write as eps or png}
}
\value{
None
}
\description{
Function to plot an aggregated radial status of EcoRisk values [0-1]
for the different sub-categories to file
}
