% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ecorisk.R
\name{plot_overtime_to_screen}
\alias{plot_overtime_to_screen}
\title{Plot timeline of EcoRisk variables to screen}
\usage{
plot_overtime_to_screen(
  data,
  timerange,
  yrange = c(0, 1),
  leg_yes = TRUE,
  leg_only = FALSE,
  varnames = NULL
)
}
\arguments{
\item{data}{EcoRisk data array
c(4/19\verb{[biomes]},8/10\verb{[nEcoRiskcomponents]},3\verb{[min,mean,max]},timeslices)}

\item{timerange}{of the data input}

\item{yrange}{range for y axis default c(0,1)}

\item{leg_yes}{plot legend (default TRUE)}

\item{leg_only}{plot only the legend? default: FALSE}

\item{varnames}{manual override of ecorisk subvariable names - default: NULL}
}
\description{
Function to plot timeline of EcoRisk variables to screen
}
\examples{
\dontrun{
plot_biome_internal_distribution(
  data = biomes,
  file = "./biomes.png"
)
}

}
