% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioCol.R
\name{plotBioCol}
\alias{plotBioCol}
\title{Plot absolute BioCol, overtime, maps, and npp into given folder}
\usage{
plotBioCol(
  biocolData,
  outFol,
  plotyears,
  minVal,
  maxVal,
  legendpos,
  startyr,
  mapyear,
  mapyear_buffer = 5,
  highlightyear,
  eps = FALSE
)
}
\arguments{
\item{biocolData}{biocol data list object (returned from calcBioCol) containing
biocol, npp_eco_overtime, npp_act_overtime, npp_pot_overtime,
npp_bioenergy_overtime, biocol_overtime, npp_harv_overtime,
biocol_overtime_perc_piref, biocol_perc, biocol_perc_piref, npp
all in GtC}

\item{outFol}{folder to write into}

\item{plotyears}{range of years to plot over time}

\item{minVal}{y-axis minimum value for plot over time}

\item{maxVal}{y-axis maximum value for plot over time}

\item{legendpos}{position of legend}

\item{startyr}{first year of biocolData object}

\item{mapyear}{year to plot biocol map for}

\item{mapyear_buffer}{+- years around mapyear to average biocol
(make sure these years exist in biocolData)}

\item{highlightyear}{year(s) that should be highlighted in overtime plot}

\item{eps}{write plots as eps, instead of png (default = FALSE)}
}
\value{
none
}
\description{
Wrapper function to plot absolute biocol, overtime, maps, and npp into given folder
}
\examples{
\dontrun{
}
}
