% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioCol.R
\name{plotBioColovertime}
\alias{plotBioColovertime}
\title{Plot absolute BioCol, overtime, maps, and npp into given folder}
\usage{
plotBioColovertime(
  biocolData,
  file,
  firstyr,
  plotyrs,
  highlightyrs = 2000,
  minVal = 0,
  maxVal = 100,
  legendpos = "topleft",
  ext = FALSE,
  eps = FALSE,
  ref = "pi"
)
}
\arguments{
\item{biocolData}{biocol data list object (returned from calcBioCol) containing
 biocol, npp_eco_overtime, npp_act_overtime, npp_pot_overtime,
 npp_bioenergy_overtime, biocol_overtime, npp_harv_overtime,
 biocol_overtime_perc_piref, biocol_perc, biocol_perc_piref, npp
all in GtC}

\item{file}{character string for location/file to save plot to}

\item{firstyr}{first year of biocol object}

\item{plotyrs}{range of years to plot over time}

\item{highlightyrs}{year(s) that should be highlighted in overtime plot (default: 2000)}

\item{minVal}{y-axis minimum value for plot over time (default: 0)}

\item{maxVal}{y-axis maximum value for plot over time (default: 100)}

\item{legendpos}{position of legend (default: "topleft")}

\item{eps}{write plots as eps, instead of png (default = FALSE)}

\item{ref}{reference period for biocol ("pi" or "act"), to either use
biocolData$biocol_overtime_perc_piref or biocolData$biocol_overtime}
}
\value{
none
}
\description{
Plot to file a comparison over time of global sums of BioCol, NPPpot, NPPeco,
and NPPact, with legend similar to Krausmann et al. 2013
}
\examples{
\dontrun{
}
}
