% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EcoRisk.R
\name{plotBiomeAveragesToScreen}
\alias{plotBiomeAveragesToScreen}
\title{Plot radial EcoRisk plot to file with 4/16 biomes}
\usage{
plotBiomeAveragesToScreen(
  data,
  biome_class_names,
  title = "",
  titleSize = 2,
  leg_scale = 0.5,
  palette = NULL
)
}
\arguments{
\item{data}{input data with dimension c(nbiome_classes,3) -- Q10,Q50,Q90 each}

\item{biome_class_names}{to write into}

\item{title}{character string title for plot, default empty}

\item{titleSize}{character string title for plot}

\item{leg_scale}{character string title for plot}

\item{palette}{color palette to plot EcoRisk with, defaults to the Ostberg 
color scheme white-blue-yellow-red}
}
\value{
None
}
\description{
Function to plot an aggregated radial status of EcoRisk values [0-1]
for the different sub-categories to file
}
\examples{
\dontrun{

}
}
