% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECO.R
\name{plotBiomeInternalDistributionToScreen}
\alias{plotBiomeInternalDistributionToScreen}
\title{Plot distribution of similarity within biomes}
\usage{
plotBiomeInternalDistributionToScreen(
  data,
  biomes_abbrv,
  title = "",
  legendtitle = "",
  scale = 1
)
}
\arguments{
\item{data}{data object with distibution - as returned by 
calculateWithInBiomeDiffs for each subcategory of meco.
dim: c(biomes,bins)}

\item{biomes_abbrv}{to mask the focusBiome from}

\item{title}{character string title for plot, default empty}

\item{legendtitle}{character string legend title, default empty}

\item{scale}{scaling factor for distribution. defaults to 1}
}
\value{
None
}
\description{
Function to plot the distribution of similarity within biomes
}
\examples{
\dontrun{

}
}
