% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECO.R, R/MECO_sav.R
\name{plotBiomes}
\alias{plotBiomes}
\title{Plot biomes to file}
\usage{
plotBiomes(
  biome_ids,
  biomeNameLength = 1,
  orderLegend = "plants",
  file,
  title = "",
  titleSize = 2,
  legYes = T,
  leg_scale = 1,
  eps = FALSE
)

plotBiomes(
  biome_ids,
  biomeNameLength = 1,
  orderLegend = "plants",
  file,
  title = "",
  titleSize = 2,
  legYes = T,
  leg_scale = 1,
  eps = FALSE
)
}
\arguments{
\item{biome_ids}{biome id as given by classify_biomes}

\item{biomeNameLength}{length of biome names in legend: 1 - abbreviation, 
2 - short name, 3 - full biome name}

\item{orderLegend}{legend order: either "plants" to first have forests, then 
grasslands, then tundra ..., or "zones" to go from north to south
(default: "plants")}

\item{file}{to write into}

\item{title}{character string title for plot, default empty}

\item{titleSize}{size of title in cex units (defaukt: 2)}

\item{legYes}{whether to plot legend (default: True)}

\item{leg_scale}{size of legend in cex units (default 0.5)}

\item{eps}{write as eps, replacing png in filename (default: True)}
}
\value{
None

None
}
\description{
Function to plot biome classification to file

Function to plot biome classification to file
}
\examples{
\dontrun{

}
\dontrun{

}
}
