% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECO.R
\name{plotBiomesAverage}
\alias{plotBiomesAverage}
\title{Plot radial gamma plot to file with 4/16 biomes}
\usage{
plotBiomesAverage(
  data,
  file,
  biome_class_names,
  title,
  titleSize = 2,
  leg_scale = 1,
  eps = FALSE
)
}
\arguments{
\item{data}{gamma data array c(4[biomes],[ngammacomponents],3[min,median,max])}

\item{file}{to write into}

\item{biome_class_names}{to write into}

\item{title}{character string title for plot}

\item{titleSize}{character string title for plot}

\item{leg_scale}{character string title for plot}

\item{eps}{write as eps or png}
}
\value{
None
}
\description{
Function to plot an aggregated radial status of gamma values [0-1]
for the different sub-categories to file
}
\examples{
\dontrun{

}
}
