% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECO.R
\name{plotBiomesAveragesToScreen}
\alias{plotBiomesAveragesToScreen}
\title{Plot radial gamma plot to file with 4/16 biomes}
\usage{
plotBiomesAveragesToScreen(
  data,
  biome_class_names,
  title,
  titleSize = 2,
  leg_scale = 0.5
)
}
\arguments{
\item{data}{input data with dimension c(nbiome_classes,3) -- Q10,Q50,Q90 each}

\item{biome_class_names}{to write into}

\item{title}{character string title for plot}

\item{titleSize}{character string title for plot}

\item{leg_scale}{character string title for plot}
}
\value{
None
}
\description{
Function to plot an aggregated radial status of gamma values [0-1]
for the different sub-categories to file
}
\examples{
\dontrun{

}
}
