% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECO.R, R/MECO_sav.R
\name{plotBiomesToScreen}
\alias{plotBiomesToScreen}
\title{Plot biomes}
\usage{
plotBiomesToScreen(
  biome_ids,
  biomeNameLength = 1,
  orderLegend = "plants",
  title = "",
  titleSize = 2,
  legYes = T,
  leg_scale = 0.5
)

plotBiomesToScreen(
  biome_ids,
  biomeNameLength = 1,
  orderLegend = "plants",
  title = "",
  titleSize = 2,
  legYes = T,
  leg_scale = 0.5
)
}
\arguments{
\item{biome_ids}{biome id as given by classify_biomes}

\item{biomeNameLength}{length of biome names in legend: 1 - abbreviation, 
2 - short name, 3 - full biome name}

\item{title}{character string title for plot, default empty}

\item{titleSize}{size of title in cex units (defaukt: 2)}

\item{legYes}{whether to plot legend (default: True)}

\item{leg_scale}{size of legend in cex units (default 0.5)}

\item{order}{legend order: either "plants" to first have forests, then 
grasslands, then tundra ..., or "zones" to go from north to south
(default: "plants")}
}
\value{
None

None
}
\description{
Function to plot biome classification

Function to plot biome classification
}
\examples{
\dontrun{

}
\dontrun{

}
}
