% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EcoRisk.R
\name{plotEcoRiskmapToScreen}
\alias{plotEcoRiskmapToScreen}
\title{Plot EcoRisk map to screen}
\usage{
plotEcoRiskmapToScreen(
  data,
  focusBiome = NULL,
  biome_classes = NULL,
  title = "",
  legendtitle = "",
  titleSize = 1,
  legYes = T,
  palette = NULL
)
}
\arguments{
\item{data}{folder of reference run}

\item{focusBiome}{highlight the biome with this id and desaturate all other (default NULL -- no highlight)}

\item{biome_classes}{to mask the focusBiome from}

\item{title}{character string title for plot, default empty}

\item{legendtitle}{character string legend title}

\item{legYes}{logical. whether to plot legend or not. defaults to TRUE}

\item{palette}{color palette to plot EcoRisk with, defaults to the Ostberg 
color scheme white-blue-yellow-red}

\item{legScale}{scaling factor for legend. defaults to 1}
}
\value{
None
}
\description{
Function to plot a global map of EcoRisk values [0-1] per grid cell to screen
}
\examples{
\dontrun{

}
}
