% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECO.R
\name{plotEcoRiskradialToScreen}
\alias{plotEcoRiskradialToScreen}
\title{Plot radial EcoRisk plot to screen}
\usage{
plotEcoRiskradialToScreen(
  data,
  title = "",
  zoom = 1,
  type = "regular",
  titleSize = 2,
  titleline = -2,
  quantile = T
)
}
\arguments{
\item{data}{EcoRisk data array c(4/19[biomes],[nEcoRiskcomponents],3[min,mean,max])}

\item{title}{character string title for plot, default empty}

\item{zoom}{scaling factor for circle plot. defaults to 1}

\item{type}{plot type, 'legend1' for variable and color legend,
'legend2' for value legend, or 'regular' (default setting)
for the regular EcoRisk plot}

\item{titleSize}{scaling factor for tile. defaults to 1}
}
\value{
None
}
\description{
Function to plot an aggregated radial status of EcoRisk values [0-1]
for the different sub-categories to screen
}
\examples{
\dontrun{

}
}
