% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECO.R
\name{plotGammaCrossTable}
\alias{plotGammaCrossTable}
\title{Plot crosstable to file showing (dis-)similarity between average biome pixels}
\usage{
plotGammaCrossTable(data, file, lmar = 3, eps = FALSE)
}
\arguments{
\item{data}{crosstable data as array with [nbiomes,nbiomes] and row/colnames}

\item{file}{to write into}

\item{lmar}{left margin for plot in lines (default: 3)}

\item{eps}{write as eps or png}
}
\value{
None
}
\description{
Function to plot to file a crosstable showing (dis-)similarity between average
biome pixels based on M-ECO (former Gamma) metric from LPJmL simulations
}
\examples{
\dontrun{

}
}
