% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECO.R
\name{plotGammaMapToScreen}
\alias{plotGammaMapToScreen}
\title{Plot gamma map to screen}
\usage{
plotGammaMapToScreen(
  data,
  focusBiome = NULL,
  biome_classes = NULL,
  title,
  legendtitle,
  titleSize = 1,
  legYes = T
)
}
\arguments{
\item{data}{folder of reference run}

\item{focusBiome}{highlight the biome with this id and desaturate all other (default NULL -- no highlight)}

\item{biome_classes}{to mask the focusBiome from}

\item{title}{character string title for plot}

\item{legendtitle}{character string legend title}

\item{legYes}{logical. whether to plot legend or not. defaults to TRUE}

\item{legScale}{scaling factor for legend. defaults to 1}
}
\value{
None
}
\description{
Function to plot a global map of gamma values [0-1] per grid cell to screen
}
\examples{
\dontrun{

}
}
