% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECO.R
\name{plotGammaRadial}
\alias{plotGammaRadial}
\title{Plot radial gamma plot to file}
\usage{
plotGammaRadial(data, file, title, legYes = T, eps = FALSE, quantile = T)
}
\arguments{
\item{data}{gamma data array c([ngammacomponents],3[min,median,max])}

\item{file}{to write into}

\item{title}{character string title for plot}

\item{legYes}{logical. whether to plot legend or not. defaults to TRUE}

\item{eps}{write as eps or png}

\item{type}{plot type, 'legend1' for variable and color legend,
'legend2' for value legend, or 'regular' (default setting)
for the regular gamma plot}
}
\value{
None
}
\description{
Function to plot an aggregated radial status of gamma values [0-1]
for the different sub-categories to file
}
\examples{
\dontrun{

}
}
