% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECO.R
\name{plotGammaRadial4}
\alias{plotGammaRadial4}
\title{Plot radial gamma plot to file with 4/16 biomes}
\usage{
plotGammaRadial4(data, biomeNames, file, eps = FALSE, quantile = T)
}
\arguments{
\item{data}{gamma data array c(4[biomes],[ngammacomponents],3[min,median,max])}

\item{file}{to write into}

\item{eps}{write as eps or png}

\item{title}{character string title for plot}
}
\value{
None
}
\description{
Function to plot an aggregated radial status of gamma values [0-1]
for the different sub-categories to file
}
\examples{
\dontrun{

}
}
