% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECO.R
\name{plotGammaRadialToScreen}
\alias{plotGammaRadialToScreen}
\title{Plot radial gamma plot to screen}
\usage{
plotGammaRadialToScreen(
  data,
  title,
  zoom = 1,
  type = "regular",
  titleSize = 2,
  titleline = -2,
  quantile = T
)
}
\arguments{
\item{data}{gamma data array c([ngammacomponents],3[min,median,max])}

\item{title}{character string title for plot}

\item{zoom}{scaling factor for circle plot. defaults to 1}

\item{type}{plot type, 'legend1' for variable and color legend,
'legend2' for value legend, or 'regular' (default setting)
for the regular gamma plot}

\item{titleSize}{scaling factor for tile. defaults to 1}
}
\value{
None
}
\description{
Function to plot an aggregated radial status of gamma values [0-1]
for the different sub-categories to screen
}
\examples{
\dontrun{

}
}
