% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGlobal.R
\name{plotGlobal}
\alias{plotGlobal}
\title{Plot global LPJmL array}
\usage{
plotGlobal(
  data,
  file,
  title = "",
  pow2max = NULL,
  pow2min = NULL,
  min = NULL,
  max = NULL,
  colPos = "GnBu",
  type = "exp",
  colNeg = "YlOrRd",
  legendtitle = "",
  legYes = TRUE,
  onlyPos = FALSE,
  eps = FALSE
)
}
\arguments{
\item{data}{array with data to plot in LPJmL specific array c(67420)}

\item{file}{character string for location/file to save plot to}

\item{title}{character string title for plot}

\item{pow2max}{for exponential legend: upper (positive) end of data range to 
plot (2^pow2max)}

\item{pow2min}{for exponential legend: smallest positive number to be 
distinguished from 0 (2^-pow2min)}

\item{min}{for linear legend: lower end of data range to plot (0 is placed 
symmetrically between min and max, if onlypos = FALSE)}

\item{max}{for linear legend: upper end of data range to plot (0 is placed 
symmetrically between min and max, if onlypos = FALSE)}

\item{colPos}{color palette for the positives}

\item{type}{string indicating whether to plot exponential (exp) or 
linear (lin) legend (default: exp)}

\item{colNeg}{color palette for the negatives}

\item{legendtitle}{character string legend title}

\item{legYes}{boolean whether to show legend (default: TRUE)}

\item{onlyPos}{boolean to show only positive half of legend (default: FALSE)}

\item{eps}{boolean whether to write eps file instead of PNG (default: FALSE)}
}
\value{
None
}
\description{
Creates a PNG/eps with a plot of a global LPJmL array
   Data is plotted in range: c(-2^pow2max,-2^-pow2min,0,2^-pow2min,2^pow2max)
   colors for pos and neg values can be given, default is Blues for the positive
   and Reds for the negative numbers
   0-range (from 2^-pow2min to 2^pow2min) is white.
   The negatives can be omitted by setting onlyPos=T, in case there are only pos values.
}
\examples{
plotGlobal(data=irrigation2006,file=paste("~/","mwateramount_2005_06.png",sep=""),
            title = paste("irrigation amount 2006 in mm/yr",sep=""),pow2max=15,pow2min=0,
            legendtitle="legendtitle",legYes=TRUE,eps=FALSE)

}
