% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCOL.R
\name{plotMCOL}
\alias{plotMCOL}
\title{Plot absolute MCOL, overtime, maps, and npp into given folder}
\usage{
plotMCOL(
  mcolData,
  outFol,
  plotyears,
  minVal,
  maxVal,
  legendpos,
  startyr,
  mapyear,
  mapyear_buffer = 5,
  highlightyear,
  eps = FALSE
)
}
\arguments{
\item{mcolData}{mcol data list object (returned from calcMCOL) containing
mcol, npp_eco_overtime, npp_act_overtime, npp_pot_overtime,
npp_bioenergy_overtime, mcol_overtime, npp_harv_overtime,
mcol_overtime_perc_piref, mcol_perc, mcol_perc_piref, npp
all in GtC}

\item{outFol}{folder to write into}

\item{plotyears}{range of years to plot over time}

\item{minVal}{y-axis minimum value for plot over time}

\item{maxVal}{y-axis maximum value for plot over time}

\item{legendpos}{position of legend}

\item{startyr}{first year of mcolData object}

\item{mapyear}{year to plot mcol map for}

\item{mapyear_buffer}{+- years around mapyear to average mcol
(make sure these years exist in mcolData)}

\item{highlightyear}{year(s) that should be highlighted in overtime plot}

\item{eps}{write plots as eps, instead of png (default = FALSE)}
}
\value{
none
}
\description{
Wrapper function to plot absolute mcol, overtime, maps, and npp into given folder
}
\examples{
\dontrun{
}
}
