% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCOL.R, R/MCOL_sav.R
\name{plotMCOLmap}
\alias{plotMCOLmap}
\title{Plot global map of MCOL to file}
\usage{
plotMCOLmap(
  data,
  file,
  title = "",
  legendtitle = "",
  zeroThreshold = 0.1,
  eps = FALSE
)

plotMCOLmap(
  data,
  file,
  title = "",
  legendtitle = "",
  zeroThreshold = 0.1,
  eps = FALSE
)
}
\arguments{
\item{data}{array containing MCOL percentage value for each gridcell}

\item{file}{character string for location/file to save plot to}

\item{title}{character string title for plot}

\item{legendtitle}{character string legend title}

\item{zeroThreshold}{smallest value to be distinguished from 0 in legend,
both for negative and positive values (default: 0.1)}

\item{eps}{write eps file instead of PNG (boolean) - (default: FALSE)}

\item{plotyears}{range of years to plot over time}
}
\value{
none

none
}
\description{
Plot global map of MCOL to file with legend colors similar to Haberl et al. 2007

Plot global map of MCOL to file with legend colors similar to Haberl et al. 2007
}
\examples{
\dontrun{
}
\dontrun{
}
}
