% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCOL.R
\name{plotMCOLovertime}
\alias{plotMCOLovertime}
\title{Plot absolute MCOL, overtime, maps, and npp into given folder}
\usage{
plotMCOLovertime(
  mcolData,
  file,
  firstyr,
  plotyrs,
  highlightyrs = 2000,
  minVal = 0,
  maxVal = 100,
  legendpos = "topleft",
  ext = FALSE,
  eps = FALSE,
  ref = "pi"
)
}
\arguments{
\item{mcolData}{mcol data list object (returned from calcMCOL) containing
 mcol, npp_eco_overtime, npp_act_overtime, npp_pot_overtime,
 npp_bioenergy_overtime, mcol_overtime, npp_harv_overtime,
 mcol_overtime_perc_piref, mcol_perc, mcol_perc_piref, npp
all in GtC}

\item{file}{character string for location/file to save plot to}

\item{firstyr}{first year of mcol object}

\item{plotyrs}{range of years to plot over time}

\item{highlightyrs}{year(s) that should be highlighted in overtime plot (default: 2000)}

\item{minVal}{y-axis minimum value for plot over time (default: 0)}

\item{maxVal}{y-axis maximum value for plot over time (default: 100)}

\item{legendpos}{position of legend (default: "topleft")}

\item{eps}{write plots as eps, instead of png (default = FALSE)}

\item{ref}{reference period for mcol ("pi" or "act"), to either use
mcolData$mcol_overtime_perc_piref or mcolData$mcol_overtime}
}
\value{
none
}
\description{
Plot to file a comparison over time of global sums of MCOL, NPPpot, NPPeco,
and NPPact, with legend similar to Krausmann et al. 2013
}
\examples{
\dontrun{
}
}
