% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECO.R
\name{plotMECOcrossTableToScreen}
\alias{plotMECOcrossTableToScreen}
\title{Plot crosstable showing (dis-)similarity between average biome pixels}
\usage{
plotMECOcrossTableToScreen(data, lmar = 3)
}
\arguments{
\item{data}{crosstable data as array with [nbiomes,nbiomes] and row/colnames}

\item{lmar}{left margin for plot in lines (default: 3)}
}
\value{
None
}
\description{
Function to plot a crosstable showing (dis-)similarity between average
biome pixels based on M-ECO (former gamma) metric from LPJmL simulations
}
\examples{
\dontrun{

}
}
