% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECO.R, R/MECO_sav.R
\name{plotMECOmap}
\alias{plotMECOmap}
\title{Plot MECO map to file}
\usage{
plotMECOmap(
  data,
  file,
  focusBiome = NULL,
  biome_classes = NULL,
  title = "",
  legendtitle = "",
  eps = FALSE,
  titleSize = 1,
  legYes = T
)

plotMECOmap(
  data,
  file,
  focusBiome = NULL,
  biome_classes = NULL,
  title = "",
  legendtitle = "",
  eps = FALSE,
  titleSize = 1,
  legYes = T
)
}
\arguments{
\item{data}{folder of reference run}

\item{file}{to write into}

\item{focusBiome}{highlight the biome with this id and desaturate all other (default NULL -- no highlight)}

\item{biome_classes}{to mask the focusBiome from}

\item{title}{character string title for plot, default empty}

\item{legendtitle}{character string legend title}

\item{eps}{write as eps or png}

\item{legYes}{logical. whether to plot legend or not. defaults to TRUE}

\item{legScale}{scaling factor for legend. defaults to 1}
}
\value{
None

None
}
\description{
Function to plot a global map of MECO values [0-1] per grid cell to file

Function to plot a global map of MECO values [0-1] per grid cell to file
}
\examples{
\dontrun{

}
\dontrun{

}
}
