% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECO.R
\name{plotMECOovertimePanel}
\alias{plotMECOovertimePanel}
\title{Plot timeline of MECO variables as panel to file with 4/16 biomes}
\usage{
plotMECOovertimePanel(
  data,
  biomeNames,
  file,
  yrange = c(0, 1),
  timerange,
  eps = FALSE,
  varnames = NULL
)
}
\arguments{
\item{data}{MECO data array c(4/19[biomes],[nMECOcomponents],3[min,mean,max])}

\item{biomeNames}{names of biomes}

\item{file}{to write into}

\item{yrange}{range for y axis (default c(0,1))}

\item{timerange}{of the data input}

\item{eps}{write as eps or png}
}
\value{
None
}
\description{
Function to plot a panel of 4/16 timelines per biome aggregated MECO values [0-1]
to file
}
\examples{
\dontrun{

}
}
