% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECO.R
\name{plotMECOradial}
\alias{plotMECOradial}
\title{Plot radial MECO plot to file}
\usage{
plotMECOradial(data, file, title = "", legYes = T, eps = FALSE, quantile = T)
}
\arguments{
\item{data}{MECO data array c(4/19[biomes],[nMECOcomponents],3[min,mean,max])}

\item{file}{to write into}

\item{title}{character string title for plot, default empty}

\item{legYes}{logical. whether to plot legend or not. defaults to TRUE}

\item{eps}{write as eps or png}

\item{type}{plot type, 'legend1' for variable and color legend,
'legend2' for value legend, or 'regular' (default setting)
for the regular MECO plot}
}
\value{
None
}
\description{
Function to plot an aggregated radial status of MECO values [0-1]
for the different sub-categories to file
}
\examples{
\dontrun{

}
}
