% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECO.R
\name{plotMECOradialPanel}
\alias{plotMECOradialPanel}
\title{Plot radial MECO panel to file with 4/16 biomes}
\usage{
plotMECOradialPanel(data, biomeNames, file, quantile = T, eps = FALSE)
}
\arguments{
\item{data}{MECO data array c(4/19[biomes],[nMECOcomponents],3[min,mean,max])}

\item{biomeNames}{names of biomes}

\item{file}{to write into}

\item{quantile}{is it quantiles or minmeanmax data? - text for whiskers}

\item{eps}{write as eps or png}
}
\value{
None
}
\description{
Function to plot an aggregated radial status of MECO values [0-1]
for the different sub-categories to file
}
\examples{
\dontrun{

}
}
