% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECO.R, R/MECO_sav.R
\name{plotOvertimeToScreen}
\alias{plotOvertimeToScreen}
\title{Plot timeline of MECO variables to screen}
\usage{
plotOvertimeToScreen(
  data,
  timerange,
  yrange = c(0, 1),
  legYes = T,
  legOnly = F,
  varnames = NULL
)

plotOvertimeToScreen(
  data,
  timerange,
  yrange = c(0, 1),
  legYes = T,
  legOnly = F,
  varnames = NULL
)
}
\arguments{
\item{data}{MECO data array 
c(4/19[biomes],8/10[nMECOcomponents],3[min,mean,max],timeslices)}

\item{timerange}{of the data input}

\item{yrange}{range for y axis default c(0,1)}

\item{legYes}{plot legend (default T)}
}
\value{
None

None
}
\description{
Function to plot timeline of MECO variables to screen

Function to plot timeline of MECO variables to screen
}
\examples{
\dontrun{

}
\dontrun{

}
}
