% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCOL.R, R/MCOL_sav.R
\name{readMCOLdata}
\alias{readMCOLdata}
\title{Calculate MCOL based on a PNV run and LU run of LPJmL}
\usage{
readMCOLdata(
  inFol_lu,
  inFol_pnv,
  startyr,
  stopyr,
  gridbased = T,
  pftbands = 11,
  cftbands = 32,
  headersize = 0,
  saveDataFile = FALSE,
  dataFile,
  ncells = 67420,
  fileType = "clm",
  varnames
)

readMCOLdata(
  inFol_lu,
  inFol_pnv,
  startyr,
  stopyr,
  gridbased = T,
  pftbands = 11,
  cftbands = 32,
  headersize = 0,
  saveDataFile = FALSE,
  dataFile,
  ncells = 67420,
  fileType = "clm",
  varnames
)
}
\arguments{
\item{inFol_lu}{folder of landuse scenario run}

\item{inFol_pnv}{folder of pnv reference run}

\item{startyr}{first year of simulations}

\item{stopyr}{last year of simulations}

\item{gridbased}{logical are pft outputs gridbased or pft-based?}

\item{pftbands}{number of natural plant functional types (== bands in fpc - 1)}

\item{cftbands}{number of crop functional types}

\item{headersize}{headersize of the output files (default 0)}

\item{saveDataFile}{whether to save input data to file (default T)}

\item{dataFile}{file to save read in input data to}

\item{ncells}{number of cells in lpjml grid}

\item{fileType}{type of output files - one of "clm", "nc", "meta" (default: "clm")}

\item{varnames}{data.frame with names (outname) and timestep of output files
see code of calcMCOL() for an example with default values}
}
\value{
list data object containing the data to compute MCOL as arrays of 
        ynpp_potential, ynpp, pftnpp_cft, pftnpp_nat, pftnpp_grasslands, 
        pftnpp_bioenergy, harvest_cft, rharvest_cft, fire, timber, cftfrac, 
        fpc, harvest_grasslands, harvest_bioenergy

list data object containing the data to compute MCOL as arrays of 
        ynpp_potential, ynpp, pftnpp_cft, pftnpp_nat, pftnpp_grasslands, 
        pftnpp_bioenergy, harvest_cft, rharvest_cft, fire, timber, cftfrac, 
        fpc, harvest_grasslands, harvest_bioenergy
}
\description{
Function to calculate MCOL based on a PNV run and LU run of LPJmL

Function to calculate MCOL based on a PNV run and LU run of LPJmL
}
\examples{
\dontrun{
}
\dontrun{
}
}
