% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECO.R
\name{readMECOData}
\alias{readMECOData}
\title{Read in output data from LPJmL to calculate the ecosystem change metric MECO}
\usage{
readMECOData(
  files_reference,
  files_scenario,
  saveFile = NULL,
  export = FALSE,
  time_span_reference,
  time_span_scenario,
  nitrogen
)
}
\arguments{
\item{files_reference}{folder of reference run}

\item{files_scenario}{folder of scenario run}

\item{saveFile}{file to save read in data to (default NULL)}

\item{export}{flag whether to export réad in data to global environment (default F)}

\item{time_span_reference}{vector of years to use as scenario period}

\item{time_span_scenario}{vector of years to use as scenario period}

\item{nitrogen}{include nitrogen outputs for pools and fluxes into MECO calculation (default F)}
}
\value{
list data object containing arrays of state_ref, mean_state_ref,
        state_scen, mean_state_scen, fpc_ref, fpc_scen, bft_ref, bft_scen,
        cft_ref, cft_scen, lat, lon, cellArea
}
\description{
Utility function to read in output data from LPJmL for calculation of MECO
}
\examples{
\dontrun{
}
}
