% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecorisk.R
\name{replace_ref_data_with_average_ref_biome_cell}
\alias{replace_ref_data_with_average_ref_biome_cell}
\title{Create modified EcoRisk data file}
\usage{
replace_ref_data_with_average_ref_biome_cell(
  data_file_in,
  data_file_out,
  biome_classes_in,
  ref_biom
)
}
\arguments{
\item{data_file_in}{path to input data}

\item{data_file_out}{path to save modified data to}

\item{biome_classes_in}{biome classes object as returned from classify_biomes}

\item{ref_biom}{reference biome from biome classes that all cells should
be compared to}
}
\description{
Function to create a modified EcoRisk data file where each reference cell is
compared to the average reference biome cell. The scenario period is
overwritten with the original reference period and all reference cells are
set to the average cell of the prescribed reference biome ref_biom
}
