% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EcoRisk.R
\name{replaceRefDataWithAverageRefBiomeCell}
\alias{replaceRefDataWithAverageRefBiomeCell}
\title{Create modified EcoRisk data file}
\usage{
replaceRefDataWithAverageRefBiomeCell(
  dataFileIn,
  dataFileOut,
  biome_classes_in,
  refBiom
)
}
\arguments{
\item{dataFileIn}{path to input data}

\item{dataFileOut}{path to save modified data to}

\item{biome_classes_in}{biome classes object as returned from classify_biomes}

\item{refBiom}{reference biome from biome classes that all cells should 
be compared to}
}
\description{
Function to create a modified EcoRisk data file where each reference cell is 
compared to the average reference biome cell. The scenario period is
overwritten with the original reference period and all reference cells are 
set to the average cell of the prescribed reference biome refBiom
}
\examples{
\dontrun{
}
}
