% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{roc_plot_paper}
\alias{roc_plot_paper}
\title{Plot ROC curve}
\usage{
roc_plot_paper(filename = NULL, values)
}
\arguments{
\item{filename}{path to plot roc curve to. default: NULL -> plot to screen}

\item{values}{roc data object as obtained from calc_roc_data(), can be an array
with dimensions = c(thresholds, internal_metric(s),
external_indicator(s), 3\link{"TP","FP","slope"})}
}
\description{
Plot ROC curve(s) - paper version
}
