% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecorisk.R
\name{state_diff_global}
\alias{state_diff_global}
\title{c based on Heyder 2011 eq. 10-13}
\usage{
state_diff_global(ref, scen, cell_area, epsilon = 10^-4)
}
\arguments{
\item{ref}{mean reference state vector of dimension c(ncells,variables)}

\item{scen}{mean scenario state vector of dimension c(ncells,variables)}

\item{cell_area}{area of each cell as a vector of dim=c(ncells)}

\item{epsilon}{threshold for variables to be treated as 0}
}
\value{
the length of the difference vector for each cell
}
\description{
c based on Heyder 2011 eq. 10-13
}
