% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecorisk.R
\name{state_diff_local}
\alias{state_diff_local}
\title{based on Heyder 2011 eq. 6-9; epsilon case handling from code
  by Sebastian Ostberg (not documented in papers)}
\usage{
state_diff_local(ref, scen, epsilon = 10^-4)
}
\arguments{
\item{ref}{mean reference state vector of dimension c(ncells,variables)}

\item{scen}{mean scenario state vector of dimension c(ncells,variables)}

\item{epsilon}{threshold for variables to be treated as 0}
}
\value{
the length of the difference vector for each cell
}
\description{
based on Heyder 2011 eq. 6-9; epsilon case handling from code
  by Sebastian Ostberg (not documented in papers)
}
