% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCOL.R, R/MCOL_sav.R
\name{calcMCOL}
\alias{calcMCOL}
\title{Calculate MCOL}
\usage{
calcMCOL(
  inFol_lu,
  inFol_pnv,
  startyr,
  stopyr,
  gridbased = T,
  npp_threshold = 20,
  pftbands = 11,
  cftbands = 32,
  dataFile = NULL,
  ncells = 67420,
  saveDataFile = FALSE,
  readPreviouslySavedData = FALSE,
  fileType = "clm",
  headersize = 0,
  varnames = NULL,
  grass_scaling = T,
  grass_harvest_file = "/p/projects/open/Fabian/LPJbox/grazing_data.RData"
)

calcMCOL(
  inFol_lu,
  inFol_pnv,
  startyr,
  stopyr,
  gridbased = T,
  npp_threshold = 20,
  pftbands = 11,
  cftbands = 32,
  dataFile = NULL,
  ncells = 67420,
  saveDataFile = FALSE,
  readPreviouslySavedData = FALSE,
  fileType = "clm",
  headersize = 0,
  varnames = NULL,
  grass_scaling = T,
  grass_harvest_file = "/p/projects/open/Fabian/LPJbox/grazing_data.RData"
)
}
\arguments{
\item{inFol_lu}{folder of landuse scenario run}

\item{inFol_pnv}{folder of pnv reference run}

\item{startyr}{first year of simulations}

\item{stopyr}{last year of simulations}

\item{gridbased}{logical are pft outputs gridbased or pft-based?}

\item{npp_threshold}{lower threshold for npp (to mask out non-lu areas
according to Haberl et al. 2007). Below MCOL will be set to 0.
(default: 20 gC/m2)}

\item{pftbands}{number of natural plant functional types (== bands in fpc - 1)}

\item{cftbands}{number of crop functional types}

\item{dataFile}{file to save/read input data to/from (default NULL)}

\item{ncells}{number of cells in lpjml grid}

\item{saveDataFile}{whether to save input data to file (default FALSE)}

\item{readPreviouslySavedData}{flag whether to read previously saved data
instead of reading it in from output files (default FALSE)}

\item{fileType}{type of output files - one of "clm", "nc" (default: "clm")}

\item{headersize}{headersize of the output files (default 0)}

\item{varnames}{data.frame with names (outname) and timestep of output files
-- can be specified to account for variable file names
(default NULL -- standard names are used)}

\item{grass_scaling}{whether to scale pasture harvest according to
data given via grass_harvest_file (default T)}

\item{grass_harvest_file}{file containing grazing data to rescale the
grassland harvests according to Herrero et al. 2013. File contains:
grazing_data list object with $name and $id of 29 world regions, and
$Herrero_2000_kgDM_by_region containing for each of these regions and
mapping_lpj67420_to_grazing_regions array with a mapping between 67420
LPJmL cells and the 29 regions}
}
\value{
list data object containing MCOL and components as arrays: mcol, 
        mcol_overtime, mcol_overtime_perc_piref, mcol_perc, mcol_perc_piref, 
        ynpp_potential, npp_act_overtime, npp_pot_overtime, npp_eco_overtime, 
        harvest_cft_overtime, npp_luc_overtime, rharvest_cft_overtime, 
        fire_overtime, timber_harvest_overtime, harvest_cft, mcol_harvest,
        grassland_scaling_factor_cellwise,  mcol_luc, mcol_luc_piref

list data object containing MCOL and components as arrays: mcol, 
        mcol_overtime, mcol_overtime_perc_piref, mcol_perc, mcol_perc_piref, 
        ynpp_potential, npp_act_overtime, npp_pot_overtime, npp_eco_overtime, 
        harvest_cft_overtime, npp_luc_overtime, rharvest_cft_overtime, 
        fire_overtime, timber_harvest_overtime, harvest_cft, mcol_harvest,
        grassland_scaling_factor_cellwise,  mcol_luc, mcol_luc_piref
}
\description{
Function to calculate MCOL

Function to calculate MCOL
}
\examples{
\dontrun{
}
\dontrun{
}
}
