% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECO.R
\name{get_biome_names}
\alias{get_biome_names}
\title{Get biome names}
\usage{
get_biome_names(biomeNameLength = 2)
}
\arguments{
\item{biomeNameLength}{integer chose from 1,2,3 for abbreviated, short, 
or full biome names}
}
\description{
Returns biome names with variable length (abbreviated, short, or full)
}
\examples{
\dontrun{
}
}
