% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_outputs.R
\name{list_outputs}
\alias{list_outputs}
\title{List required LPJmL outputs and temporal resolution}
\usage{
list_outputs(metric = "all", only_first_filename = TRUE)
}
\arguments{
\item{metric}{Character string containing name of metric to get
required outputs. Available options are `c("meco", "mcol", "biome",
"nitrogen", "lsc", "bluewater", "greenwater", "water", "biosphere")`
or just `"all"` or `"benchmark"`.}

\item{only_first_filename}{Logical. If TRUE, only the first file name will be
returned for each output. If FALSE, all file names will be returned.}
}
\value{
List of output IDs with required resolution and
        file names for a given metric
}
\description{
Function to return a list of output IDs with required resolution and
file names for a given metric. The list is based on the metric_files.yml file
in the boundaries package (`"./inst/metric_files.yml"`).
}
\examples{
\dontrun{

}
}
