% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_outputs.R, R/utils.R
\name{list_outputs}
\alias{list_outputs}
\title{List required LPJmL outputs and temporal resolution}
\usage{
list_outputs(metric = "all", only_first_filename = TRUE)

list_outputs(metric = "all", only_first_filename = TRUE)
}
\arguments{
\item{metric}{character string/list of strings. metrics to list outputs for
can be one of:
"all" - list all outputs for all metrics
"ecorisk" - list outputs for ecorisk metric without nitrogen
"ecorisk_nitrogen" - list outputs for ecorisk metric with nitrogen
"biocol" - list outputs for biocol metric
"biome" - list outputs for the biome classification}

\item{only_first_filename}{if several legal output names are listed, only
output the first of them (default: TRUE)}
}
\value{
List of output IDs with required resolution and
        file names for a given metric

list object with required outputs, their required temporal resolution
and if it is optional
}
\description{
Function to return a list of output IDs with required resolution and
file names for a given metric. The list is based on the metric_files.yml file
in the boundaries package (`"./inst/metric_files.yml"`).

List required output files for given metric based on parameter file
\code{inst/extfiles/metric_files.yml}
}
\examples{
\dontrun{

}
\dontrun{
list_outputs(metric = "ecorisk_nitrogen")
}

}
