% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_biocol.R
\name{plot_biocol}
\alias{plot_biocol}
\title{Plot absolute BioCol, overtime, maps, and npp into given folder}
\usage{
plot_biocol(
  biocol_data,
  path_write,
  plotyears,
  min_val,
  max_val,
  legendpos,
  details = FALSE,
  start_year,
  mapyear,
  mapyear_buffer = 5,
  highlightyear,
  eps = FALSE
)
}
\arguments{
\item{biocol_data}{biocol data list object (returned from calc_biocol)
containing biocol_overtime, biocol_overtime_abs,
biocol_overtime_abs_frac_piref, biocol_overtime_frac_piref,
biocol_overtime_frac, biocol_overtime_abs_frac, npp_harv_overtime,
npp_luc_overtime, npp_act_overtime, npp_pot_overtime,
npp_eco_overtime, harvest_grasslands_overtime,
harvest_bioenergy_overtime, harvest_cft_overtime,
rharvest_cft_overtime, fire_overtime, timber_harvest_overtime,
wood_harvest_overtime, biocol, biocol_frac, npp, biocol_frac_piref,
npp_potential, npp_ref, harvest_cft, rharvest_cft, biocol_harvest,
biocol_luc, lat, lon, cellarea}

\item{path_write}{folder to write outputs into}

\item{plotyears}{range of years to plot over time}

\item{min_val}{y-axis minimum value for plot over time}

\item{max_val}{y-axis maximum value for plot over time}

\item{legendpos}{position of legend}

\item{details}{show all harvest components or not}

\item{start_year}{first year of biocol_data object}

\item{mapyear}{year to plot biocol map for}

\item{mapyear_buffer}{+- years around mapyear to average biocol
(make sure these years exist in biocol_data) - default: 5}

\item{highlightyear}{year(s) that should be highlighted in overtime plot}

\item{eps}{write plots as eps, instead of png (default = FALSE)}
}
\description{
Wrapper function to plot absolute biocol, overtime, maps, and npp into given
folder
}
\examples{
\dontrun{
plot_biocol(
   biocol_data = biocol,
   path_write = "~/BioCol_plots/",
   plotyears = c(1980,2014),
   min_val = 0,
   max_val = 90,
   legendpos = "left",
   start_year = 1980,
   mapyear = 2000,
   highlightyear = 2000,
   eps = FALSE
)
}

}
