% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_biocol.R
\name{plot_biocol_map}
\alias{plot_biocol_map}
\title{Plot global map of BioCol to file}
\usage{
plot_biocol_map(
  data,
  lon,
  lat,
  file = NULL,
  title = "",
  legendtitle = "",
  zero_threshold = 0.001,
  haberl_legend = FALSE,
  eps = FALSE
)
}
\arguments{
\item{data}{array containing BioCol percentage value for each gridcell}

\item{lon}{longitude array}

\item{lat}{latitude array}

\item{file}{to write into, if not supplied (default is NULL) write to screen}

\item{title}{character string title for plot (default: "")}

\item{legendtitle}{character string legend title (default: "")}

\item{zero_threshold}{smallest value to be distinguished from 0 in legend,
both for negative and positive values (default: 0.001)}

\item{haberl_legend}{use color palette from Haberl et al.? (default: FALSE)}

\item{eps}{write eps file instead of PNG (boolean) - (default: FALSE)}
}
\description{
Plot global map of BioCol to file with legend colors similar to
Haberl et al. 2007
}
\examples{
\dontrun{
plot_biocol_map(
  data = biocol$biocol_frac[, "2000"] * 100,
  file = "./BioCol_map_yr2000.png",
)
}

}
