% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ecorisk.R
\name{plot_biome_averages}
\alias{plot_biome_averages}
\title{Plot radial EcoRisk plot to file with 4/16 biomes}
\usage{
plot_biome_averages(
  data,
  file,
  biome_class_names,
  title = "",
  title_size = 2,
  leg_scale = 1,
  eps = FALSE,
  palette = NULL
)
}
\arguments{
\item{data}{EcoRisk data array c(4[biomes],[nEcoRiskcomponents],
3[min,median,max])}

\item{file}{to write into}

\item{biome_class_names}{to write into}

\item{title}{character string title for plot, default empty}

\item{title_size}{character string title for plot}

\item{leg_scale}{character string title for plot}

\item{eps}{write as eps, replacing png in filename (default: True)}

\item{palette}{color palette to plot EcoRisk with, defaults to the Ostberg
color scheme white-blue-yellow-red}
}
\value{
None
}
\description{
Function to plot an aggregated radial status of EcoRisk values [0-1]
for the different sub-categories to file
}
