% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_biomes.R
\name{plot_biomes}
\alias{plot_biomes}
\title{Plot global distribution of lpjml simulated biomes}
\usage{
plot_biomes(
  biome_data,
  file_name = NULL,
  display_area = FALSE,
  cellarea = NULL,
  order_legend = c(1, 2, 9, 10, 11, 3, 4, 5, 12, 13, 14, 6, 7, 8, 15, 16, 17, 18, 19),
  projection = "+proj=robin",
  grid_path = NULL
)
}
\arguments{
\item{biome_data}{output (list) from classify_biomes()}

\item{file_name}{directory for saving the plot (character string)}

\item{display_area}{boolean, adding occupied area per biome (default F)}

\item{cellarea}{array with cellarea in m2 per gridcell (default NULL)}

\item{order_legend}{in which order the biomes should be displayed
default: c(1,2,9,10,11,3,4,5,12,13,14,6,7,8,15,16,17,18,19)}

\item{projection}{character string defining the projection,
default: "+proj=robin", for mercator: "+proj=wgs84"}

\item{grid_path}{character string providing the path to a grid file}
}
\description{
Plots a map with the biome distribution as derived from a lpjml run based
on the "classify_biomes" function
}
\examples{
\dontrun{

biomes <- classify_biomes(
 config_reference = path_reference,
 time_span_reference = as.character(2008:2017),
 savanna_proxy = list(vegc = 7500)
)

plot_biomes(
  biome_data = biomes,
  file_name = "/p/projects/open/Johanna/R/biomes.pfd"
  grid_path = ".grid.bin.json"
)
}

}
