% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecorisk.R
\name{plot_ecorisk_cross_table}
\alias{plot_ecorisk_cross_table}
\title{Plot crosstable to file showing (dis-)similarity between average biome pixels}
\usage{
plot_ecorisk_cross_table(data, file, lmar = 3, eps = FALSE, palette = NULL)
}
\arguments{
\item{data}{crosstable data as array with [nbiomes,nbiomes] and row/colnames}

\item{file}{to write into}

\item{lmar}{left margin for plot in lines (default: 3)}

\item{eps}{write as eps or png}

\item{palette}{color palette to plot EcoRisk with, defaults to the Ostberg
color scheme white-blue-yellow-red}
}
\value{
None
}
\description{
Function to plot to file a crosstable showing (dis-)similarity between
average biome pixels based on EcoRisk (former Gamma) metric from LPJmL
simulations
}
