% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ecorisk.R
\name{plot_ecorisk_map}
\alias{plot_ecorisk_map}
\title{Plot EcoRisk map to file}
\usage{
plot_ecorisk_map(
  ecorisk_object,
  plot_dimension,
  file = NULL,
  focus_biome = NULL,
  biome_classes = NULL,
  title = "",
  legendtitle = "",
  eps = FALSE,
  title_size = 1,
  leg_yes = TRUE,
  palette = NULL
)
}
\arguments{
\item{ecorisk_object}{ecorisk object from which to plot}

\item{plot_dimension}{which dimension from ecorisk object to plot}

\item{file}{to write into, if not supplied (default is NULL) write to screen}

\item{focus_biome}{highlight the biome with this id and desaturate all other
(default NULL -- no highlight)}

\item{biome_classes}{to mask the focus_biome from}

\item{title}{character string title for plot, default empty}

\item{legendtitle}{character string legend title}

\item{eps}{write as eps or png}

\item{leg_yes}{logical. whether to plot legend or not. defaults to TRUE}

\item{palette}{color palette to plot EcoRisk with, defaults to the Ostberg
color scheme white-blue-yellow-red}

\item{leg_scale}{scaling factor for legend. defaults to 1}
}
\description{
Function to plot a global map of EcoRisk values \link{0-1} per grid cell to file
}
\examples{
\dontrun{
plot_biome_internal_distribution(
  data = biomes,
  file = "./biomes.png"
)
}

}
