% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ecorisk.R
\name{plot_ecorisk_over_time_panel}
\alias{plot_ecorisk_over_time_panel}
\title{Plot timeline of EcoRisk variables as panel to file with 4/16 biomes}
\usage{
plot_ecorisk_over_time_panel(
  data,
  biome_names,
  file = NULL,
  yrange = c(0, 1),
  timerange,
  eps = FALSE,
  varnames = NULL
)
}
\arguments{
\item{data}{EcoRisk data array c(4/19\verb{[biomes]},\verb{[nEcoRiskcomponents]},
3\verb{[min,mean,max]})}

\item{biome_names}{names of biomes}

\item{file}{to write into (if not supplied - default NULL - prints to screen)}

\item{yrange}{range for y axis (default c(0,1))}

\item{timerange}{of the data input}

\item{eps}{write as eps or png}

\item{varnames}{list vector with variable names}
}
\description{
Function to plot a panel of 4/16 timelines per biome aggregated EcoRisk
values \verb{[0-1]}
to file
}
\examples{
\dontrun{
plot_biome_internal_distribution(
  data = biomes,
  file = "./biomes.png"
)
}

}
