% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ecorisk.R
\name{plot_ecorisk_radial}
\alias{plot_ecorisk_radial}
\title{Plot radial EcoRisk plot to file}
\usage{
plot_ecorisk_radial(
  data,
  file,
  title = "",
  leg_yes = TRUE,
  eps = FALSE,
  use_quantile = TRUE
)
}
\arguments{
\item{data}{EcoRisk data array c(4/19\verb{[biomes]},\verb{[nEcoRiskcomponents]},
3\verb{[min,mean,max]})}

\item{file}{to write into}

\item{title}{character string title for plot, default empty}

\item{leg_yes}{logical. whether to plot legend or not. defaults to TRUE}

\item{eps}{write as eps or png}

\item{use_quantile}{show quantiles or min,mean,max}
}
\description{
Function to plot an aggregated radial status of EcoRisk values \verb{[0-1]}
for the different sub-categories to file
}
\examples{
\dontrun{
plot_biome_internal_distribution(
  data = biomes,
  file = "./biomes.png"
)
}

}
