% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ecorisk.R
\name{plot_ecorisk_radial_panel}
\alias{plot_ecorisk_radial_panel}
\title{Plot radial EcoRisk panel to file with 4/16 biomes}
\usage{
plot_ecorisk_radial_panel(
  data,
  biome_names,
  file = NULL,
  use_quantile = TRUE,
  eps = FALSE
)
}
\arguments{
\item{data}{EcoRisk data array c(4/19\verb{[biomes]},\verb{[nEcoRiskcomponents]},
3\verb{[min,mean,max]})}

\item{biome_names}{names of biomes}

\item{file}{to write into (if not supplied - default NULL - prints to screen)}

\item{use_quantile}{is it quantiles or minmeanmax data? - text for whiskers}

\item{eps}{write as eps or png}
}
\description{
Function to plot an aggregated radial status of EcoRisk values \verb{[0-1]}
for the different sub-categories to file
}
\examples{
\dontrun{
plot_ecorisk_radial_panel(
  data = ecorisk_disaggregated_full[-c(3, 18, 19), c(1:5, 8, 9, 13), ],
  biome_names = biomes,
  file = "./biomes.png"
)
}

}
