% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ecorisk.R
\name{plot_ecorisk_radial_to_screen}
\alias{plot_ecorisk_radial_to_screen}
\title{Plot radial EcoRisk plot to screen}
\usage{
plot_ecorisk_radial_to_screen(
  data,
  title = "",
  zoom = 1,
  type = "regular",
  title_size = 2,
  titleline = -2,
  use_quantile = TRUE
)
}
\arguments{
\item{data}{EcoRisk data array c(\verb{[nEcoRiskcomponents]},
3\verb{[min,mean,max]})}

\item{title}{character string title for plot, default empty}

\item{zoom}{scaling factor for circle plot. defaults to 1}

\item{type}{plot type, 'legend1' for variable and color legend,
'legend2' for value legend, or 'regular' (default setting)
for the regular EcoRisk plot}

\item{title_size}{scaling factor for tile. defaults to 1}

\item{titleline}{line at which the title will be displayed. defaults to -2}

\item{use_quantile}{use quantiles or min,mean,max. defaults to TRUE}
}
\description{
Function to plot an aggregated radial status of EcoRisk values \verb{[0-1]}
for the different sub-categories to screen
}
\examples{
\dontrun{
plot_biome_internal_distribution(
  data = biomes,
  file = "./biomes.png"
)
}

}
