% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_global.R
\name{plot_global}
\alias{plot_global}
\title{Plot global LPJmL array}
\usage{
plot_global(
  data,
  file = NULL,
  title = "",
  pow2min = NULL,
  pow2max = NULL,
  min = NULL,
  max = NULL,
  brks = NULL,
  palette = NULL,
  col_pos = "GnBu",
  type = "exp",
  col_neg = "YlOrRd",
  legendtitle = "",
  leg_yes = TRUE,
  only_pos = FALSE,
  n_legend_ticks = 20,
  min_0 = 0.01,
  eps = FALSE
)
}
\arguments{
\item{data}{array with data to plot in LPJmL specific array c(67420)}

\item{file}{character string for location/file to save plot to, if not
supplied, the plot is displayed to screen (default: NULL)}

\item{title}{character string title for plot}

\item{pow2min}{for exponential legend: smallest positive number to be
distinguished from 0 (2^-pow2min)}

\item{pow2max}{for exponential legend: upper (positive) end of data range to
plot (2^pow2max)}

\item{min}{for linear legend: lower end of data range to plot (0 is placed
symmetrically between min and max, if onlypos = FALSE)}

\item{max}{for linear legend: upper end of data range to plot (0 is placed
symmetrically between min and max, if onlypos = FALSE)}

\item{col_pos}{color palette for the positives}

\item{type}{string indicating whether to plot manual (man),
exponential (exp) or linear (lin) legend (default: exp).
man requires: parameters brks and palette defined,
exp requires: parameters pow2min and pow2max given defined,
lin requires: parameters min and max defined}

\item{col_neg}{color palette for the negatives}

\item{legendtitle}{character string legend title}

\item{leg_yes}{boolean whether to show legend (default: TRUE)}

\item{only_pos}{boolean to show only positive half of legend (default: FALSE)}

\item{n_legend_ticks}{(default: 20)}

\item{min_0}{(default: 0.01)}

\item{eps}{boolean whether to write eps file instead of PNG (default: FALSE)}
}
\description{
Plot global LPJmL array to file (if file argument is given) or screen (else).
type argument controls plot type: (exponential, linear, or manual legend).
Depending on this more parameters are required.
Data plot ranges:
exp: c(-2^pow2max,-2^-pow2min,0,2^-pow2min,2^pow2max)
lin: c(min,max)
man: brks
colors for pos and neg values can be given, default is Blues for the
positive and Reds for the negative numbers 0-range (from 2^-pow2min to
2^pow2min) is white.
The negatives can be omitted by setting only_pos=TRUE, in case there are
only pos values.
}
\examples{
\dontrun{
plot_global(
  data = biocol_data$biocol[,"2015"]),
  file = "BioCol_absolute_2015.png",
  type = "exp",
  pow2min = 0,
  pow2max = 12,
  legendtitle = "GtC",
  leg_yes = TRUE,
  only_pos = FALSE,
  eps = FALSE
)
}

}
