% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECO.R
\name{plotBiomesToScreen}
\alias{plotBiomesToScreen}
\title{Plot radial gamma plot to file with 4/16 biomes}
\usage{
plotBiomesToScreen(
  biome_ids,
  biome_class_names,
  title,
  titleSize = 2,
  legYes = T,
  leg_scale = 0.5
)
}
\arguments{
\item{title}{character string title for plot}

\item{data}{gamma data array c(4[biomes],[ngammacomponents],3[min,median,max])}

\item{file}{to write into}

\item{eps}{write as eps or png}
}
\value{
None
}
\description{
Function to plot an aggregated radial status of gamma values [0-1]
for the different sub-categories to file
}
\examples{
\dontrun{

}
}
